base functional Flask
